+++ includes/lib/fpdf/fpdf.php
@@ -1,3 +1,0 @@
-<?php
-// Dummy FPDF file for PHPStan
-class FPDF {}