+++ includes/lib/fpdi/src/autoload.php
@@ -1,2 +1,0 @@
-<?php
-// Dummy FPDI autoload for PHPStan